Fix g++ warnings

+++ fem/dfem/autopa.hpp
@@ -47,7 +47,7 @@ class NewAutoActionCallback
    const int trial_vdim;
    const int total_trial_op_dim;
    const int num_qp;
-   Array<size_t> &dependent_inputs_trial_op_dim;
+   Array<int> &dependent_inputs_trial_op_dim;
    const int num_dependent_inputs;
    const ElementDofOrdering element_dof_ordering;
    const int q1d_;
@@ -85,7 +85,7 @@ public:
                          const int trial_vdim,
                          const int total_trial_op_dim,
                          const int num_qp,
-                         Array<size_t> &dependent_inputs_trial_op_dim,
+                         Array<int> &dependent_inputs_trial_op_dim,
                          const int num_dependent_inputs,
                          const ElementDofOrdering element_dof_ordering,
                          const int q1d,
@@ -164,7 +164,7 @@ public:
                                 const int trial_vdim,
                                 const int total_trial_op_dim,
                                 const int num_qp,
-                                Array<size_t> &dependent_inputs_trial_op_dim,
+                                Array<int> &dependent_inputs_trial_op_dim,
                                 const int num_dependent_inputs,
                                 const ElementDofOrdering &ordering,
                                 const ThreadBlocks &thread_blocks,
@@ -279,7 +279,7 @@ public:
                      for (int k = 0; k < test_op_dim; k++)
                      {
                         real_t sum = 0.0;
-                        size_t m_offset = 0;
+                        int m_offset = 0;
                         for (int s_i = 0; s_i < num_dependent_inputs; s_i++)
                         {
                            const int s = dpitod(s_i, 0);
@@ -287,7 +287,7 @@ public:
                            auto d_qp = Reshape(&(shadow_shmem[s])[0], trial_vdim, trial_op_dim, num_qp);
                            for (int j = 0; j < trial_vdim; j++)
                            {
-                              for (size_t m = 0; m < trial_op_dim; m++)
+                              for (int m = 0; m < trial_op_dim; m++)
                               {
                                  sum += qpdc(i, k, j, m + m_offset, q, e) * d_qp(j, m, q);
                               }
